[Agent Builder] fix workflow tools returning empty output
